Changed type and status fields to small int

+++ backend/tasks/models.py
@@ -9,7 +9,7 @@ class Tile(models.Model):
         ARCHIVED = 3
 
     launch_date = models.DateField(default=date.today)
-    status = models.IntegerField(choices=Status.choices)
+    status = models.SmallIntegerField(choices=Status.choices)
 
     def __str__(self):
         return f'{self.pk} - {self.status}'
@@ -24,7 +24,7 @@ class Task(models.Model):
     title = models.CharField(max_length=100)
     order = models.PositiveSmallIntegerField()
     description = models.TextField()
-    type = models.IntegerField(choices=Type.choices)
+    type = models.SmallIntegerField(choices=Type.choices)
     tile = models.ForeignKey(Tile, related_name="tasks", on_delete=models.CASCADE)
 
     def __str__(self):
